% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_dictionary}
\alias{data_dictionary}
\title{Data dictionary with old values and standardized values}
\format{
Dataframe with three columns and 10 rows
\describe{
\item{old_name}{The original variable name before standadization and cleanup}
\item{new_name}{The new standardized variable name}
\item{description}{Description of the data in a given variable}
}
}
\usage{
data_dictionary
}
\description{
This data dictionary demonstrated how data dictionaries can be used to clean variable names in R
}
\examples{
data(data_dictionary)
}
\keyword{datasets}
