% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{class}
\name{dataset-class}
\alias{dataset-class}
\alias{dataset}
\title{A \code{dataset} is an Object  Identified by a Name and a List of
  Features}
\description{
Each instance of this class represents a daset identified by a
  list of features and a name. Always use \code{\link{dataset.new}} to
  instantiate this class, never the direct constructor.
}
\section{Slots}{

\describe{
\item{\code{name}}{the name of this dataset}

\item{\code{features}}{a list of name-value pairs denoting the features of the
dataset}

\item{\code{data}}{the actual data, a list, which will be known under the name
\code{name} and which is annotated with \code{features}}
}}

