% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset.new}
\alias{dataset.new}
\title{Create an Instance of \code{\link{dataset}}}
\usage{
dataset.new(name, features, data)
}
\arguments{
\item{name}{the name of this dataset}

\item{features}{a list of name-value pairs denoting the features of the dataset}

\item{data}{the list of with the data, which is named by \code{name} and annotated with \code{features}.}
}
\description{
Instantiate \code{\link{dataset}} using this method.
}
