% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asDataFrame.R
\name{datasets.as.data.frame}
\alias{datasets.as.data.frame}
\title{Convert the \code{\link{dataset}} Features, Feature Names, and
  \code{\link{dataset}} Names into a \code{\link{data.frame}}}
\usage{
datasets.as.data.frame(datasets, stringsAsFactors = TRUE)
}
\arguments{
\item{datasets}{the \code{\link{dataset}s}}

\item{stringsAsFactors}{convert strings to factors if \code{TRUE}, keep them
as strings if \code{FALSE}}
}
\value{
the data frame
}
\description{
Create a data frame which represents all the meta-information
  about the \code{\link{dataset}s}.
}
