% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asLists.R
\name{datasets.as.lists}
\alias{datasets.as.lists}
\title{Transform a List of Datasets to a List with Named Values}
\usage{
datasets.as.lists(data, groups, transformers)
}
\arguments{
\item{data}{the list of \code{\link{dataset}}s}

\item{groups}{the group names}

\item{transformers}{a list of functions whose names will be the names of the
new features and the new feature values are computed by these functions}
}
\value{
a list of lists, where each item belongs to exactly one group
}
\description{
A list of \code{\link{dataset}}s is translated to another list,
  where each element is, in turn, a list. Each such list belongs to one group
  and, as first element, has the group name. This name is followed by other
  features computed by the functions in list \code{transformers} and named
  accordingly.
}
