% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchLoad.R
\name{datasets.batchLoad}
\alias{datasets.batchLoad}
\title{Load the a Sequence of \code{\link{dataset}}s}
\usage{
datasets.batchLoad(path = getwd(),
  selector = path.extensionRegExp(extensions = "csv"),
  dataLoader = function(file) read.csv(file),
  featuresFolder = file.path(path, "../features"),
  featuresLoader = datasets.feature.load.default,
  namer = datasets.names.namer, check.directory = NULL, cores = 1L,
  logging = (cores <= 1L))
}
\arguments{
\item{path}{the source folder}

\item{selector}{the selector for the files}

\item{dataLoader}{a \code{function(file)} invoked for every single file
selected by the selector which returns one data item to be attached to the
\code{data} member of a dataset}

\item{featuresFolder}{the folder where to look for feature information,
\code{NULL} if no feature information is needed}

\item{featuresLoader}{a function accepting two parameters,
\code{featureFolder} and \code{components} and returning a named list of
features, or \code{NULL} if no features are needed. see
\code{\link{datasets.feature.load.default}} for documentation}

\item{namer}{receives the name components, i.e., the directory names and can
add, delete, or modify elements of the vector the name}

\item{check.directory}{a function which can choose if a directory should be
followed or not}

\item{cores}{the number of cores to be used for loading}

\item{logging}{should progress information be printed: either \code{TRUE} for
printing to the console via \code{\link{print}}, \code{FALSE} for no
logging, or a path to a file receiving logging information}
}
\value{
a list of \code{\link{dataset}s} instances
}
\description{
Load all the \code{\link{dataset}}s from a folder structure.
}
