% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featuresDefaultLoader.R
\name{datasets.feature.load.default}
\alias{datasets.feature.load.default}
\title{The Default Loader for Features}
\usage{
datasets.feature.load.default(featureFolder, components, values.eval = TRUE)
}
\arguments{
\item{featureFolder}{the folder with the features}

\item{components}{the components}

\item{values.eval}{the evaluated values}
}
\value{
a named list, where the names are the feature names and the values
  the corresponding feature values
}
\description{
The default loader for the \code{\link{dataset}} features is the
  \code{\link{features.txtLoader}}
}
