% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featuresTxtLoader.R
\name{datasets.feature.load.text}
\alias{datasets.feature.load.text}
\title{The Text-based Loader for Features}
\usage{
datasets.feature.load.text(featureFolder, components, values.eval = TRUE)
}
\arguments{
\item{featureFolder}{the folder with the features}

\item{components}{the components}

\item{values.eval}{the evaluated values}
}
\value{
a named list, where the names are the feature names and the values
  the corresponding feature values
}
\description{
The text-based loader for the \code{\link{dataset}} features
  searches the \code{featureFolder} for all text files whose names match to
  the supplied path \code{components}. Each such file is loaded as a
  Java-style properties file with \code{name=value} type data and
  \code{#}-based comments. The resulting list contains all the corresponding
  names and values. Values can be parsed if \code{value.eval} is set to
  \code{TRUE}.
}
