% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureNames.R
\name{datasets.featureNames}
\alias{datasets.featureNames}
\title{Get the Names of All Features of All Datasets}
\usage{
datasets.featureNames(datasets)
}
\arguments{
\item{datasets}{the list of \code{\link{dataset}s} from where we want to
extract the feature names}
}
\value{
the feature names (alphabetically sorted)
}
\description{
Get the names of all features of all \code{\link{dataset}s}
}
