% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featuresUnique.R
\name{datasets.features.unique}
\alias{datasets.features.unique}
\title{Find all the Unique Combinations of Feature Values and the
  Corresponding Datasets}
\usage{
datasets.features.unique(datasets, features.ignore = NULL,
  features.keep = NULL)
}
\arguments{
\item{datasets}{the list of instances of \code{\link{dataset}}}

\item{features.ignore}{a vector of names of features to ignore, or
\code{NULL} if no features should explicitly be ignored}

\item{features.keep}{a vector of the names of the (only) features to
consider, or \code{NULL} if all non-ignored features should be considered}
}
\value{
a list of two components, \code{features} and \code{selection}. The
  \code{features} component contains the lists of unique feature value
  combinations. The \code{selection} component contains, for each element of
  \code{features}, a vector with indexes into the input \code{datasets} list
  where elements with these features are located
}
\description{
Search the features of a set of \code{\link{dataset}s} for
  unique combinations of values of the specified features.
}
