% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureValues.R
\name{datasets.featureValues}
\alias{datasets.featureValues}
\title{Get a Vector with the Values of one Feature for all
  \code{\link{dataset}s}}
\usage{
datasets.featureValues(datasets, featureName)
}
\arguments{
\item{datasets}{the list of \code{\link{dataset}s} from where we want to
extract the feature names}

\item{featureName}{the name of the feature}
}
\value{
the vector with the values of this feature
}
\description{
For each \code{\link{dataset}}, get the value of the specified
  feature. If the feature is undefined, its value is \code{NA}.
}
