% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeByFeatures.R
\name{datasets.merge.by.features}
\alias{datasets.merge.by.features}
\title{Create a Combination or Selection of \code{\link{dataset}s}}
\usage{
datasets.merge.by.features(datasets, features.ignore = NULL,
  features.keep = NULL, data.name = datasets.names.join,
  data.merge = .data.merge, data.create = dataset.new)
}
\arguments{
\item{datasets}{the list of instances of \code{\link{dataset}}}

\item{features.ignore}{a vector of names of features to delete, or
\code{NULL} if no features should explicitly be deleted}

\item{features.keep}{a vector of the names of the (only) features to keep, or
\code{NULL} if all non-deleted features should be kept}

\item{data.name}{a function used to join names of datasets that now have
identical features, see \code{\link{datasets.names.join}} for documentation}

\item{data.merge}{a function for merging the \code{data} lists; by default
\code{unlist(x, recursive=FALSE)}}

\item{data.create}{a function for creating new datasets, by default
\code{\link{dataset.new}}}
}
\value{
a list of two components, \code{datasets} and \code{selection}. The
  \code{datasets} component contains instances of \code{\link{dataset}} where
  all names have been transformed according to \code{data.name} and the features
  have been selected according to \code{features.ignore} and
  \code{features.keep} where the \code{data} members of instances with the
  same features have been merged. The \code{selection} component contains,
  for each element of \code{datasets}, a vector with indexes into the input
  \code{datasets} list from where its components stem
}
\description{
Merge a set of \code{\link{dataset}s} according to a given name transformation.
}
