% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeBySelection.R
\name{datasets.merge.by.selection}
\alias{datasets.merge.by.selection}
\title{Create a Combination or Selection of \code{\link{dataset}s}}
\usage{
datasets.merge.by.selection(datasets, selection = seq_along(datasets),
  data.name = dataset.joinNames, data.merge = .data.merge,
  data.create = dataset.new)
}
\arguments{
\item{datasets}{the list of instances of \code{\link{dataset}s}}

\item{selection}{a list of vectors of model indexes. \code{\link{dataset}s}
in the same index are joined}

\item{data.name}{a function used to join names of datasets that now have
identical features, see \code{\link{datasets.joinNames}} for documentation}

\item{data.merge}{a function for merging the \code{data} lists; by default
\code{unlist(x, recursive=FALSE)}}

\item{data.create}{a function for creating new datasets, by default
\code{\link{dataset.new}}}
}
\value{
a list of  instances of \code{\link{dataset}} where all names have
  been transformed according to \code{data.name}, the features have been selected
  according to selection, the data lists have been merged using
  \code{data.merge}, and the new instances have been created via
  \code{data.create}.
}
\description{
Merge a set of \code{\link{dataset}s} according to a given
  selection.
}
