% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNamesUnique.R
\name{datasets.names.make.unique}
\alias{datasets.names.make.unique}
\title{Receive a List of \code{\link{dataset}s} Instances and Make their
  Names Unique}
\usage{
datasets.names.make.unique(datasets)
}
\arguments{
\item{datasets}{the list of \code{\link{dataset}s} instances}
}
\value{
a list of \code{\link{dataset}s} instances where no name appears
  twice
}
\description{
Receive a list of instances of class \code{\link{dataset}s} and
  make sure that all of their names are unique.
}
