% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultNamer.R
\name{datasets.names.namer}
\alias{datasets.names.namer}
\title{The Default Namer for Datasets}
\usage{
datasets.names.namer(components)
}
\arguments{
\item{components}{the name components}
}
\value{
the dataset name representing the \code{components}
}
\description{
This is the default method to create names from name components.
}
