% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{datasets.names}
\alias{datasets.names}
\title{Get the Names of All \code{\link{dataset}s} in a List}
\usage{
datasets.names(datasets)
}
\arguments{
\item{datasets}{the list of  \code{\link{dataset}s}}
}
\value{
the vector of names of the  \code{\link{dataset}s}
}
\description{
This function extracts all the names of the
  \code{\link{dataset}s} in a listand returns them in the same order as the
  \code{\link{dataset}s} appear.
}
