% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectByFeatures.R
\name{datasets.select.by.features}
\alias{datasets.select.by.features}
\title{Select a Dataset by Features}
\usage{
datasets.select.by.features(datasets, features, getIndex = FALSE)
}
\arguments{
\item{datasets}{the list of \code{\link{dataset}} instances}

\item{features}{a list with named elements, we only return datasets where all features match}

\item{getIndex}{if \code{FALSE} (by default), this function returns the
datasets with the matching features; if \code{TRUE}, just the indexes of the
datasets is returned.}
}
\value{
the datasets whose features match \code{features} (\code{getIndex==FALSE})
  or their indexes (\code{getIndex==TRUE}); or an empty list if no dataset matches
  the given features
}
\description{
Return the datasets whose features match the provided features,
  or their indexes (if {@code getIndex==TRUE)}). This function returns a list
  of only the datasets whose features match to all the constraints given in
  \code{features}. Each element of \code{features} must be named and the name
  refers to a feature in the dataset. The values of the features can be
  provided as list or vector of allowed values.
}
