% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectByName.R
\name{datasets.select.by.name}
\alias{datasets.select.by.name}
\title{Select a Datasets by Name}
\usage{
datasets.select.by.name(datasets, include = NULL, exclude = list(),
  getIndex = FALSE)
}
\arguments{
\item{datasets}{the list of \code{\link{dataset}} instances}

\item{include}{the names we are looking for: if this is not \code{NULL}, then
only datasets with a name in this list will be returned}

\item{exclude}{the names we do not want: if this is not \code{NULL}, then
only datasets whose name does not appear in this list will be permitted}

\item{getIndex}{if \code{FALSE} (by default), this function returns the
datasets with the matching name; if \code{TRUE}, just the indexes of the
datasets are returned.}
}
\value{
the dataset whose names match \code{name} (\code{getIndex==FALSE}) or
  their indexes (\code{getIndex==TRUE})
}
\description{
Return the datasets whose names matches the provided
  \code{include} names while being different from the \code{exclude} names.
  If {@code getIndex==TRUE)}, return their indices instead.
}
