% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncd_data}
\alias{ncd_data}
\title{A data set for patients with Non Communicable Diseases(NCDs)}
\format{
Dataframe with 10 columns and 365 rows
\describe{
\item{patient_id}{Patient's unique identifier ranging from 1 to 365}
\item{new_name}{Patient's age in years at the time of the study ranging from 18 to 85 years}
\item{gender}{Patient's gender, which is male and females}
\item{bmi}{Patient's Body Mass Index in kg/m2}
\item{blood_pressure}{Patient's systolic blood pressure in mm/Hg}
\item{cholestrol}{Patient's cholestrol levels in mg/dL}
\item{glucose}{Patient's fasting glucose level in mg/dL}
\item{hemoglobin}{Patient's hemoglobin level in g/dL}
\item{heart_rate}{Patient's heart rate in beats per minute}
\item{diagnosis}{Patient's health condition which can be Healthy, Pre-diabetic, Diabetic and Hypertension}
}
}
\usage{
ncd_data
}
\description{
This is synthetic dataset for 365 patients with Non communicable diseases mainly hypertension and diabetes
}
\examples{
data(ncd_data)
}
\keyword{datasets}
