% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_digits.R
\name{remove_digits}
\alias{remove_digits}
\title{Remove Digits and Clean Text Strings}
\usage{
remove_digits(text_string)
}
\arguments{
\item{text_string}{A character vector containing text strings to be cleaned.}
}
\value{
A character vector with cleaned text.
}
\description{
This function removes trailing numbers and underscores from text strings,
replaces underscores and hyphens with spaces, and converts the string to title case.
}
\examples{
remove_digits("consultation_3")   # Returns "Consultation"
remove_digits("intake_1")         # Returns "Intake"
remove_digits("follow-up_2")      # Returns "Follow Up"
remove_digits("check-in-4")       # Returns "Check In"
}
