% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_rand12.R
\name{scoring_rand12_unweighted}
\alias{scoring_rand12_unweighted}
\title{Generates unweighted Physical Component Scores and Mental Component Scores from RAND-12 or RAND-36}
\usage{
scoring_rand12_unweighted(
  data,
  RAND36_1 = Health_rating_now,
  RAND36_3b = Moderate_activity4w,
  RAND36_3d = Stairs4w,
  RAND36_4b = Physical_health_lessthanwanted4w,
  RAND36_4c = Physical_health_limited4w,
  RAND36_5b = Psychological_health_lessthanwanted4w,
  RAND36_5c = Psychological_healthlimitation4w,
  RAND36_8 = Pain_limited_activities4w,
  RAND36_9d = Relaxed_feeling4w,
  RAND36_9e = Energetic_feeling4w,
  RAND36_9f = Depressed_feeling4w,
  RAND36_10 = PhysicalAndEmotional_Social4w
)
}
\arguments{
\item{data}{Data frame containing the variables needed for scoring RAND-12}

\item{RAND36_1}{The variable corresponding to RAND36_1 / RAND12_1}

\item{RAND36_3b}{The variable corresponding to RAND36_3b / RAND12_2a}

\item{RAND36_3d}{The variable corresponding to RAND36_3d / RAND12_2b}

\item{RAND36_4b}{The variable corresponding to RAND36_4b / RAND12_3a}

\item{RAND36_4c}{The variable corresponding to RAND36_4c / RAND12_3b}

\item{RAND36_5b}{The variable corresponding to RAND36_5b / RAND12_4a}

\item{RAND36_5c}{The variable corresponding to RAND36_5c / RAND12_4b}

\item{RAND36_8}{The variable corresponding to RAND36_8 / RAND12_5}

\item{RAND36_9d}{The variable corresponding to RAND36_9d / RAND12_6a}

\item{RAND36_9e}{The variable corresponding to RAND36_9e / RAND12_6b}

\item{RAND36_9f}{The variable corresponding to RAND36_9f / RAND12_6c}

\item{RAND36_10}{The variable corresponding to RAND36_10 / RAND12_7}
}
\value{
A data frame with unweighted RAND-12 PCS (PCS12_unwh) and MCS (MCS12_unwh) scores
}
\description{
This function generates the unweighted RAND-12 PCS and MCS scores that range from 0 to 100, with higher scores indicating better HRQoL.
}
\references{
Andersen, J. R., et al. (2022). Correlated physical and mental health composite scores for the RAND-36 and RAND-12 health surveys: Can we keep them simple? Health and Quality of Life Outcomes, 20(1), 89. https://doi.org/10.1186/s12955-022-01992-0
}
