% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_participant_age.R
\name{calc_age}
\alias{calc_age}
\title{Load unprocessed followup submissions}
\usage{
calc_age(socID, refDate = Sys.Date(), unit = "year", cutoff_2000 = 22L)
}
\arguments{
\item{x}{Messy email.}
}
\value{
Clean email.
}
\description{
Load unprocessed followup submissions
}
