% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_duplicates.R
\name{check_duplicates}
\alias{check_duplicates}
\title{Check for duplicates in dataset}
\usage{
check_duplicates(data, ...)
}
\arguments{
\item{...}{Variables to control for duplicates}

\item{df}{The data frame to be processed.}
}
\value{
None
}
\description{
This function checks for duplicates in a data frame. It allows you to specify the output formats and generate metadata such as codebooks and dictionaries.
}
