% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_duplicates.R
\name{check_duplicates}
\alias{check_duplicates}
\title{Check for duplicates in a dataset based on specified groups}
\usage{
check_duplicates(data, ...)
}
\arguments{
\item{data}{The input dataset to check for duplicates}

\item{...}{Unquoted variable names of the grouping variables}
}
\value{
A data frame with duplicates removed based on the specified grouping variables
}
\description{
This function checks for duplicates in a dataset based on specified grouping variables.
}
\examples{
check_duplicates(mtcars, cyl, gear)

}
