% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_codebook.R
\name{create_codebook}
\alias{create_codebook}
\title{Create a codebook for the provided data}
\usage{
create_codebook(data)
}
\arguments{
\item{data}{The dataset for which to create the codebook}
}
\value{
A data frame containing variable labels, types, and frequency distributions
}
\description{
This function generates a codebook containing variable labels, types, and frequency distributions for the input data.
}
\examples{
create_codebook(mtcars)

}
