% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_codebook.R
\name{create_codebook}
\alias{create_codebook}
\title{Load unprocessed followup submissions}
\usage{
create_codebook(data)
}
\arguments{
\item{x}{Messy email.}
}
\value{
Clean email.
}
\description{
Load unprocessed followup submissions
}
