% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dd.R
\name{create_dd}
\alias{create_dd}
\title{Create a Named Vector for Data Dictionary Mapping}
\usage{
create_dd(.data, old_name = "old_name", new_name = "new_name")
}
\arguments{
\item{.data}{A data frame containing the columns specified in \code{old_name} and \code{new_name}.}

\item{old_name}{A character string specifying the column name in \code{.data} that contains the old variable names. Default is \code{"old_name"}.}

\item{new_name}{A character string specifying the column name in \code{.data} that contains the new variable names. Default is \code{"new_name"}.}
}
\value{
A named character vector where the old names are the names and the new names are the values.
}
\description{
This function generates a named vector that maps old variable names to new names.
}
\examples{
dd <- data.frame(old_name = c("Patient ID", "Patient's Age", "Patient Sex"),
                new_name = c("ID", "age", "sex"))
create_dd(dd, "old_name", "new_name")
# Named vector: c(Patient ID = "ID", B = "age", C = "sex")
}
