% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectByFeatures.R
\name{datasets.select.by.features}
\alias{datasets.select.by.features}
\title{Select a Dataset by Features}
\usage{
datasets.select.by.features(datasets, include = list(),
  exclude = list(), getIndex = FALSE)
}
\arguments{
\item{datasets}{the list of \code{\link{dataset}} instances}

\item{include}{a list with named elements, we only return datasets where all
features match}

\item{exclude}{a list with named elements, we only return datasets where all
features match}

\item{getIndex}{if \code{FALSE} (by default), this function returns the
datasets with the matching features; if \code{TRUE}, just the indexes of
the datasets is returned.}
}
\value{
the datasets whose features match \code{features}
  (\code{getIndex==FALSE}) or their indexes (\code{getIndex==TRUE}); or an
  empty list if no dataset matches the given features
}
\description{
Return the datasets whose features match the provided feature
  restrictions, or their indexes (if {@code getIndex==TRUE)}). This function
  returns a list of only the datasets whose features match to all the
  constraints given in \code{feature.include} and \code{exclude}. Each
  element of \code{include} and \code{exclude} must be named and the name
  refers to a feature in the dataset. The values of the features can be
  provided as list or vector of allowed values. A dataset is only returned if
  \itemize{ \item For each tuple \code{a=b} in \code{include}, the feature of
  name \code{a} of the dataset has one value from \code{b}. \code{b} can
  either be a single value, a vector, or a list. \item For each tuple
  \code{c=d} in \code{exclude}, the feature of name \code{c} of the dataset
  has one value from \code{d}. \code{d} can either be a single value, a
  vector, or a list. }
}
