% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectByName.R
\name{datasets.select.by.name}
\alias{datasets.select.by.name}
\title{Select a Dataset by Name}
\usage{
datasets.select.by.name(datasets, name, getIndex = FALSE)
}
\arguments{
\item{datasets}{the list of \code{\link{dataset}} instances}

\item{name}{the name we are looking for}

\item{getIndex}{if \code{FALSE} (by default), this function returns the
dataset with the matching name; if \code{TRUE}, just the index of the
dataset is returned.}
}
\value{
the dataset whose name matches \code{name} (\code{getIndex==FALSE})
  or its index (\code{getIndex==TRUE}); or \code{NULL} if no dataset matches
  the given name
}
\description{
Return the dataset whose name matches the provided name, or its
  index (if {@code getIndex==TRUE)}).
}
