% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_long_variables.R
\name{drop_long_variables}
\alias{drop_long_variables}
\title{Drop long variables from a dataset}
\usage{
drop_long_variables(data, variable_length = 32)
}
\arguments{
\item{data}{The dataset from which to drop long variables.}

\item{variable_length}{The maximum character length for a variable name. Default is 32.}
}
\value{
A list containing the modified dataset and a message with the names of dropped variables.
}
\description{
This function drops variables from a dataset that have names longer than a specified length.
}
\examples{
data <- iris
drop_long_variables(data, 5)

}
