% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_results_for_research.R
\name{export_results_for_research}
\alias{export_results_for_research}
\title{Export Results for Research}
\usage{
export_results_for_research(
  df,
  results_directory = "results",
  results_filename = "results",
  results_date = Sys.Date(),
  write_as_csv = TRUE,
  write_as_excel = FALSE,
  write_as_qs = FALSE,
  write_as_spss = FALSE,
  write_as_stata = FALSE,
  write_as_sas = FALSE,
  write_dictionary = FALSE,
  write_codebook_word = FALSE,
  write_codebook_html = FALSE
)
}
\arguments{
\item{df}{The data frame to be processed.}

\item{results_directory}{The path and name of the results directory (optional, default is "results").}

\item{results_filename}{The name of the results file (optional, default is "results").}

\item{results_date}{The date to be appended to the directory and file names (optional, default is Sys.Date()).}

\item{write_as_csv}{A logical indicating whether to write the results as a CSV file (optional, default is TRUE).}

\item{write_as_excel}{A logical indicating whether to write the results as an Excel file (optional, default is FALSE).}

\item{write_as_qs}{A logical indicating whether to write the results as a QS file (optional, default is FALSE).}

\item{write_as_spss}{A logical indicating whether to write the results as an SPSS file (optional, default is FALSE).}

\item{write_as_stata}{A logical indicating whether to write the results as a Stata file (optional, default is FALSE).}

\item{write_as_sas}{A logical indicating whether to write the results as a SAS file (optional, default is FALSE).}

\item{write_dictionary}{A logical indicating whether to generate a dictionary for the data frame (optional, default is FALSE).}

\item{write_codebook_word}{A logical indicating whether to generate a Word document codebook for the data frame (optional, default is FALSE).}

\item{write_codebook_html}{A logical indicating whether to generate an HTML codebook table for the data frame (optional, default is FALSE).}
}
\value{
None
}
\description{
This function exports the results of a data frame and labelled data for research purposes. It allows you to specify the output formats and generate metadata such as codebooks and dictionaries.
}
\examples{
# Export results for research
export_results_for_research(df = my_data, results_directory = "output", results_filename = "my_results", write_as_csv = TRUE, write_as_excel = TRUE, write_dictionary = TRUE)

}
