% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_readme.R
\name{generate_readme}
\alias{generate_readme}
\title{Generate README File from Template}
\usage{
generate_readme(output_name = "README_template.txt", overwrite = FALSE)
}
\arguments{
\item{output_name}{Character. The name of the output README file to create. Defaults to \code{"README_template.txt"}.}

\item{overwrite}{Logical. If \code{TRUE}, allows overwriting the existing file. Defaults to \code{FALSE}.}
}
\value{
Invisibly returns the name of the created or existing README file.
}
\description{
This function creates a README file using an embedded template stored as internal data within the package.
It ensures not to overwrite an existing file unless explicitly permitted.
}
\examples{
generate_readme(output_name = "Project_Readme.txt", overwrite = FALSE)
}
