% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_column_names.R
\name{get_column_names}
\alias{get_column_names}
\title{Load unprocessed followup submissions}
\usage{
get_column_names(data)
}
\arguments{
\item{x}{Messy email.}
}
\value{
Clean email.
}
\description{
Load unprocessed followup submissions
}
