% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_participant_age.R,
%   R/get_participant_age_5int.R
\name{get_participant_age}
\alias{get_participant_age}
\title{Load unprocessed followup submissions}
\usage{
get_participant_age(
  socID,
  refDate = Sys.Date(),
  unit = "year",
  cutoff_2000 = 22L
)

get_participant_age(
  socID,
  refDate = Sys.Date(),
  unit = "year",
  cutoff_2000 = 22L
)
}
\arguments{
\item{x}{Messy email.}
}
\value{
Clean email.

Clean email.
}
\description{
Load unprocessed followup submissions

Load unprocessed followup submissions
}
