% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_participant_birthdate.R
\name{get_participant_birthdate}
\alias{get_participant_birthdate}
\title{Get participant birthdate based on social ID}
\usage{
get_participant_birthdate(
  socID,
  refDate = Sys.Date(),
  unit = "year",
  cutoff_2000 = 22L
)
}
\arguments{
\item{socID}{Social ID of the participant}

\item{refDate}{Reference date for calculations (default is current date)}

\item{unit}{The unit of time to calculate the birthdate difference in (default is "year")}

\item{cutoff_2000}{Cutoff year for two-digit year parsing (default is 22)}
}
\value{
A date object representing the participant's birthdate
}
\description{
This function calculates the birthdate of a participant based on their social ID.
}
