% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_participant_birthdate.R
\name{get_participant_birthdate}
\alias{get_participant_birthdate}
\title{Load unprocessed followup submissions}
\usage{
get_participant_birthdate(
  socID,
  refDate = Sys.Date(),
  unit = "year",
  cutoff_2000 = 22L
)
}
\arguments{
\item{x}{Messy email.}
}
\value{
Clean email.
}
\description{
Load unprocessed followup submissions
}
