% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_email.R
\name{prep_email}
\alias{prep_email}
\title{Load unprocessed followup submissions}
\usage{
prep_email(x)
}
\arguments{
\item{x}{Messy email.}
}
\value{
Clean email.
}
\description{
Load unprocessed followup submissions
}
