% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_email.R
\name{prep_email}
\alias{prep_email}
\title{Preprocess email addresses for cleaning}
\usage{
prep_email(x)
}
\arguments{
\item{x}{Character vector containing email addresses.}
}
\value{
A character vector with preprocessed email addresses.
}
\description{
This function preprocesses email addresses for cleaning by replacing common mistakes with correct domain names and removing extra spaces.
}
\examples{
prep_email("example@email.com")
}
