% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_name.R
\name{prep_name}
\alias{prep_name}
\title{Load unprocessed followup submissions}
\usage{
prep_name(x)
}
\arguments{
\item{x}{Messy email.}
}
\value{
Clean email.
}
\description{
Load unprocessed followup submissions
}
