% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_phone.R
\name{prep_phone}
\alias{prep_phone}
\title{Load unprocessed followup submissions}
\usage{
prep_phone(x)
}
\arguments{
\item{x}{Messy email.}
}
\value{
Clean email.
}
\description{
Load unprocessed followup submissions
}
