% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_SocID.R
\name{prep_SocID}
\alias{prep_SocID}
\title{Corrects the Norwegian Social Identification (SocID) number by left-padding it with zeros to have a length of 11.}
\usage{
prep_SocID(SocID)
}
\arguments{
\item{SocID}{A character representing the SocID to be corrected.}
}
\value{
A character string with the corrected SocID.
}
\description{
Corrects the Norwegian Social Identification (SocID) number by left-padding it with zeros to have a length of 11.
}
\examples{
prep_SocID("12345") # Returns "000000012345"
}
