% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_SocID.R
\name{prep_SocID}
\alias{prep_SocID}
\title{Load unprocessed followup submissions}
\usage{
prep_SocID(SocID)
}
\arguments{
\item{x}{Messy email.}
}
\value{
Clean email.
}
\description{
Load unprocessed followup submissions
}
